requirements.txt
albumentations==2.0.8
face_emotion==1.1.1
facenet_pytorch==2.6.0
Flask==3.1.1
hydra-core==1.3.2
matplotlib==3.10.3
numpy==2.3.2
omegaconf==2.3.0
opencv_python==4.12.0.88
pandas==2.3.1
Pillow==11.3.0
pytest==8.4.1
python-dotenv==1.1.1
pytorch_lightning==2.5.2
Requests==2.32.4
scikit_learn==1.7.1
setuptools==80.9.0
tensorflow==2.19.0
timm==1.0.19
torch==2.6.0
torch==2.7.1
torchmetrics==1.8.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.54.1
